+++ packages/main/cypress/specs/ComboBox.cy.tsx
@@ -1,6 +1,8 @@
 import ComboBox from "../../src/ComboBox.js";
 import ComboBoxItem from "../../src/ComboBoxItem.js";
 import ComboBoxItemGroup from "../../src/ComboBoxItemGroup.js";
+import Link from "../../src/Link.js";
+import Input from "../../src/Input.js";
 
 describe("Security", () => {
 	it("tests setting malicious text to items", () => {
@@ -87,6 +89,198 @@ describe("Keyboard interaction", () => {
 	});
 });
 
+describe("Keyboard interaction when pressing Ctrl + Alt + F8 for navigation", () => {
+	beforeEach(() => {
+		cy.mount(<>
+			<ComboBox valueState="Negative">
+				<div slot="valueStateMessage">
+					Custom error value state message with a <Link href="#">Link</Link>  <Link href="#">Second Link</Link>.
+				</div>
+				<ComboBoxItem text="alert('XSS')"></ComboBoxItem>
+				<ComboBoxItem text="<b onmouseover=alert('XSS')></b>"></ComboBoxItem>
+				<ComboBoxItem text="Albania"></ComboBoxItem>
+			</ComboBox>
+			<Input id="nextInput" class="input2auto" placeholder="Next input"></Input>
+		</>);
+	});
+
+	it("Should move the focus from the ComboBox to the first link in the value state message", () => {
+		cy.get("ui5-combobox")
+			.shadow()
+			.find("input")
+			.as("innerInput");
+
+		cy.get("ui5-combobox")
+			.as("combobox");
+
+		cy.get("@innerInput")
+			.realClick()
+			.should("be.focused");
+
+		cy.realPress(["Control", "Alt", "F8"]);
+
+		cy.get("@combobox")
+			.shadow()
+			.find("ui5-popover")
+			.as("popover")
+			.should("have.class", "ui5-valuestatemessage-popover");
+
+		cy.get("@popover")
+			.should("have.attr", "open")
+
+		cy.get("ui5-link")
+			.eq(0)
+			.should("have.focus");
+	});
+
+	it("Pressing [Tab] moves the focus to the next value state message link. Pressing [Tab] again closes the popup and moves the focus to the next input", () => {
+		cy.get("ui5-combobox")
+			.as("combobox");
+
+		cy.get("@combobox")
+			.shadow()
+			.find("input")
+			.as("innerInput");
+
+		cy.get("@innerInput")
+			.realClick()
+			.should("be.focused");
+
+		cy.realPress(["Control", "Alt", "F8"]);
+
+		cy.get("@combobox")
+			.shadow()
+			.find("ui5-popover")
+			.as("ui5-popover")
+			.should("have.attr", "open");
+
+		cy.get("ui5-link")
+			.eq(0)
+			.as("firstLink")
+			.should("have.focus");
+
+		cy.get("@firstLink")
+			.realPress("Tab");
+
+		cy.get("@firstLink")
+			.should("not.have.focus");
+
+		cy.get("ui5-link")
+			.eq(1)
+			.as("secondLink")
+			.should("have.focus");
+
+		cy.get("@secondLink")
+			.realPress("Tab");
+
+
+		cy.get("ui5-input")
+			.as("input");
+
+		cy.get("@input")
+			.should("have.focus");
+	});
+
+	it("Pressing [Shift+Tab] moves the focus from the second value state message link to the first. Pressing it again shifts the focus to the ComboBox", () => {
+		cy.get("ui5-combobox")
+			.shadow()
+			.find("input")
+			.as("innerInput");
+
+		cy.get("ui5-combobox")
+			.as("combobox");
+
+		cy.get("@innerInput")
+			.realClick()
+			.should("be.focused");
+
+		cy.realPress(["Control", "Alt", "F8"]);
+
+		cy.get("@combobox")
+			.shadow()
+			.find("ui5-popover")
+			.as("ui5-popover")
+			.should("have.attr", "open");
+
+		cy.get("ui5-link")
+			.eq(0)
+			.as("firstLink")
+			.should("have.focus");
+
+		cy.get("@firstLink")
+			.realPress("Tab");
+
+		cy.get("@firstLink")
+			.should("not.have.focus");
+
+		cy.get("ui5-link")
+			.eq(1)
+			.as("secondLink")
+			.should("have.focus");
+
+		cy.get("@secondLink")
+			.realPress(["Shift", "Tab"]);
+
+		cy.get("@firstLink")
+			.should("have.focus");
+
+		cy.get("@firstLink")
+			.realPress(["Shift", "Tab"]);
+
+		cy.get("@innerInput")
+			.should("have.focus");
+	});
+
+	it("When pressing [Down Arrow] while focused on the first value state message link and suggestions are open, the focus moves to the next suggestion item", () => {
+		cy.get("ui5-combobox")
+			.shadow()
+			.find("input")
+			.as("innerInput");
+
+		cy.get("ui5-combobox")
+			.as("combobox");
+
+		cy.get("@innerInput")
+			.realClick()
+			.should("be.focused");
+
+		cy.realPress(["Control", "Alt", "F8"]);
+
+		cy.get("@combobox")
+			.shadow()
+			.find("ui5-responsive-popover")
+			.as("popover");
+
+		cy.get("@combobox")
+			.realClick()
+			.should("be.focused");
+
+		cy.realType("A");
+
+		cy.get("@popover")
+			.should("have.attr", "open");
+
+		cy.get("@innerInput")
+			.realClick()
+			.should("be.focused");
+
+		cy.realPress(["Control", "Alt", "F8"]);
+
+		cy.get("ui5-link")
+			.as("firstLink")
+			.should("have.focus");
+
+		cy.get("@firstLink")
+			.realPress("ArrowDown");
+
+		cy.get("@combobox")
+			.should("have.attr", "value", "Albania");
+
+		cy.get("@combobox")
+			.find("[ui5-cb-item]").eq(2).should("have.prop", "focused", true);
+	});
+});
+
 describe("Event firing", () => {
 	it("tests if open and close events are fired correctly", () => {
 		cy.mount(
@@ -168,6 +362,7 @@ describe("Event firing", () => {
 		cy.get("@changeStub").should("not.have.been.called");
 	});
 });
+
 describe("Accessibility", () => {
 	it("should announce the associated label when ComboBox is focused", () => {
 		cy.mount(
@@ -194,5 +389,3 @@ describe("Accessibility", () => {
 			});
 	});
 });
-
-